+++ Source/Veno/Audio/Engine/ModModulator.h
@@ -14,7 +14,7 @@ public:
 
     ~ModModulator () override;
 
-    float getValue () override;
+    double getValue () override;
 
     void update () override;
