+++ Source/Veno/Audio/Engine/VeNoMatrix.h
@@ -13,7 +13,7 @@ struct VeNoMatrixTarget
 {
     std::string name;
     std::string source;
-    float amount = 0; // always 0 to 1 <-- apply amount to modulator
+    double amount = 0; // always 0 to 1 <-- apply amount to modulator
     [[nodiscard]] std::string toString () const;
 
     static VeNoMatrixTarget* fromString (const std::string& value);
@@ -34,7 +34,7 @@ public:
 
     void removeModulator (const std::string& name);
 
-    void setMatrixModulation (const std::string& name, const std::string& source, float amount);
+    void setMatrixModulation (const std::string& name, const std::string& source, double amount);
 
     std::unique_ptr<XmlElement> saveMatrixToXML ();
